+++ contracts/interfaces/IMessageLine.sol
@@ -4,8 +4,8 @@ pragma solidity ^0.8.0;
 interface IMessageLine {
     event MessageSent(
         bytes32 indexed _messageId,
-        uint64 _fromChainId,
-        uint64 _toChainId,
+        uint256 _fromChainId,
+        uint256 _toChainId,
         address _fromDapp,
         address _toDapp,
         bytes _message,
@@ -17,10 +17,10 @@ interface IMessageLine {
 
     event ReceiverError(bytes32 indexed _messageId, bytes _reason, address _toLine);
 
-    function send(uint64 _toChainId, address _toDapp, bytes memory _payload, bytes memory _params) external payable;
+    function send(uint256 _toChainId, address _toDapp, bytes memory _payload, bytes memory _params) external payable;
 
     function estimateFee(
-        uint64 _toChainId, // Dest lineRegistry chainId
+        uint256 _toChainId, // Dest lineRegistry chainId
         bytes calldata _payload,
         bytes calldata _params
     ) external view returns (uint256);
